Fix float64 casting

+++ mem/mem_windows.go
@@ -84,7 +84,7 @@ func SwapMemoryWithContext(ctx context.Context) (*SwapMemoryStat, error) {
 		Total:       tot,
 		Used:        used,
 		Free:        free,
-		UsedPercent: float64(used / tot),
+		UsedPercent: float64(used) / float64(tot),
 	}
 
 	return ret, nil
